requirements.txt
django
redis
aioredis
hiredis
django-redis-cache
djangorestframework
psycopg2
aiohttp>=2.0.0,<2.3.0
websockets>=3.1,<4.0
git+https://github.com/Rapptz/discord.py@rewrite
